+++ CHANGELOG.md
@@ -62,6 +62,7 @@
 - Fixed `<Markdown>` rerender on source change #2020 !3467 @AntonBallmaier
 - Fixed loading more entries in store log #2023 !3472 @AntonBallmaier
 - Fixed link in food share point map bubble #1940 !3471 @AntonBallmaier
+- Fixed store distance display for users without valid geolocation !3476 @AntonBallmaier
 
 ## Refactoring
 - Reduce complexity of RegionTransaction.GetJoinMessage() by session information !3242 @KoeMai